+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Sanity check csv rows for disallowed values
+
 ## [0.4.5] - 2023-03-18
 
 - Improved error handling and display